refactor: Split large admin and forum components into smaller, manageable pieces
Split admin.tsx (909 lines) into:
- AdminDashboard: Analytics dashboard with charts and metrics
- AdminCategoryManagement: Forum category CRUD operations
- AdminUserManagement: User role management
- AdminSettings: Platform and forum settings configuration

Split advanced-forum.tsx (798 lines) into:
- ForumStats: Forum statistics cards
- ForumTopicList: Topic listing with search/filters
- ForumTopicDetail: Single topic view with posts
- CreateTopicDialog: Topic creation dialog

Benefits:
- Improved code organization and maintainability
- Better separation of concerns
- Easier testing and debugging
- Reduced file complexity
