+++ client/src/components/admin/admin-user-management.tsx
@@ -1,0 +1,105 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
+import { Badge } from '@/components/ui/badge';
+import { useToast } from '@/hooks/use-toast';
+import { apiRequest } from '@/lib/queryClient';
+import { Users } from 'lucide-react';
+
+interface User {
+  id: number;
+  username: string;
+  email: string;
+  role: string;
+  isActive: boolean;
+  reputation: number;
+  createdAt: string;
+}
+
+interface AdminUserManagementProps {
+  users: User[];
+  isLoading: boolean;
+}
+
+export function AdminUserManagement({ users, isLoading }: AdminUserManagementProps) {
+  const { toast } = useToast();
+  const queryClient = useQueryClient();
+
+  const updateUserRoleMutation = useMutation({
+    mutationFn: async ({ userId, role }: { userId: number; role: string }) => {
+      return apiRequest(`/api/admin/users/${userId}/role`, {
+        method: 'PATCH',
+        body: JSON.stringify({ role }),
+      });
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['/api/admin/users'] });
+      toast({ title: 'User role updated successfully!' });
+    },
+    onError: () => {
+      toast({ title: 'Failed to update user role', variant: 'destructive' });
+    },
+  });
+
+  const handleUpdateUserRole = (userId: number, role: string) => {
+    updateUserRoleMutation.mutate({ userId, role });
+  };
+
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle className="flex items-center gap-2">
+          <Users className="h-5 w-5" />
+          User Management
+        </CardTitle>
+        <CardDescription>
+          Manage user roles and permissions
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <div className="space-y-4">
+          {isLoading ? (
+            <div>Loading users...</div>
+          ) : Array.isArray(users) && users.length > 0 ? (
+            users.map((user: User) => (
+              <div key={user.id} className="flex items-center justify-between p-4 border rounded-lg">
+                <div className="flex items-center gap-3">
+                  <div className="w-10 h-10 bg-primary rounded-full flex items-center justify-center text-primary-foreground font-medium">
+                    {user.username.charAt(0).toUpperCase()}
+                  </div>
+                  <div>
+                    <h3 className="font-medium">{user.username}</h3>
+                    <p className="text-sm text-muted-foreground">{user.email}</p>
+                    <p className="text-xs text-muted-foreground">
+                      Reputation: {user.reputation} • Joined: {new Date(user.createdAt).toLocaleDateString()}
+                    </p>
+                  </div>
+                </div>
+                <div className="flex items-center gap-2">
+                  <Select
+                    value={user.role}
+                    onValueChange={(role) => handleUpdateUserRole(user.id, role)}
+                  >
+                    <SelectTrigger className="w-32">
+                      <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                      <SelectItem value="user">User</SelectItem>
+                      <SelectItem value="moderator">Moderator</SelectItem>
+                      <SelectItem value="admin">Admin</SelectItem>
+                    </SelectContent>
+                  </Select>
+                  <Badge variant={user.isActive ? "default" : "secondary"}>
+                    {user.isActive ? "Active" : "Inactive"}
+                  </Badge>
+                </div>
+              </div>
+            ))
+          ) : (
+            <div className="text-center text-muted-foreground">No users found</div>
+          )}
+        </div>
+      </CardContent>
+    </Card>
+  );
+}
